Adjusted Entitiy definitions

+++ conwaywebapi/Database/BoardCell.cs
@@ -2,8 +2,8 @@ namespace ConwayWebApi.Database
 {
     public class BoardCell
     {
-        public long ID {get; set; }
-        public long BoardID { get; set; }
+        public int ID {get; set; }
+        public int BoardID { get; set; }
         public int X { get; set; }
         public int Y { get; set; }
     }
